+++ src/app/layout.tsx
@@ -1,33 +1,21 @@
-import type { Metadata } from "next";
-import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 
-const geistSans = Geist({
-  variable: "--font-geist-sans",
-  subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
-});
-
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+export const metadata = {
+  title: "The Ark of Wisdom",
+  description: "AI 中心の共同体",
 };
 
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
-    <html lang="en">
-      <body
-        className={`${geistSans.variable} ${geistMono.variable} antialiased`}
-      >
-        {children}
+    <html lang="ja">
+      <body className="bg-white text-gray-900">
+        <header className="bg-blue-500 p-4 text-white text-center">
+          <h1 className="text-2xl font-bold">The Ark of Wisdom</h1>
+        </header>
+        <main className="container mx-auto p-4">{children}</main>
+        <footer className="bg-gray-800 text-white text-center p-4 mt-10">
+          © 2025 The Ark of Wisdom
+        </footer>
       </body>
     </html>
   );
